pyproject.toml
[project]
name = "md2do"
version = "0.1.0"
description = "Speech-to-LLM todo.md updater"
requires-python = ">=3.8.1"
dependencies = [
    "openai",
    "pyyaml",
    "pygame",
    "pydub",
    "black>=23.0.0",
    "SpeechRecognition",
    "pyaudio"
]

[tool.black]
line-length = 88
